// projects/sspset/src/models/informacion-financiera.ts
export class InformacionFinanciera {
    anio: number= 0;
    pia: number= 0;
    pim: number= 0;
    dev: number= 0;
    ene: number= 0;
    feb: number= 0;
    mar: number= 0;
    abr: number= 0;
    may: number= 0;
    jun: number= 0;
    jul: number= 0;
    ago: number= 0;
    set: number= 0;
    oct: number= 0;
    nov: number= 0;
    dic: number= 0;
    compromisoanual: number= 0;
    certificadoanual: number= 0;
}